+++ src/lib.rs
@@ -179,6 +179,9 @@
 //! - `"rust_1_82"`(disabled by default):
 //! Enables formatting of additional items that require Rust 1.82.0 to do so.
 //!
+//! - `"rust_1_88"`(disabled by default):
+//! Enables formatting of additional items that require Rust 1.88.0 to do so.
+//!
 //! - `"derive"`(disabled by default):
 //! Enables the [`PanicFmt` derive] macro.
 //!
@@ -269,6 +272,9 @@ mod fmt_impls {
 
     #[cfg(feature = "non_basic")]
     mod fmt_range;
+
+    #[cfg(all(feature = "non_basic", feature = "rust_1_88"))]
+    mod rust_1_88_nonbasic_fmt_impls;
 }
 
 pub use crate::{