+++ tests/main_tests/rust_1_82_types_tests.rs
@@ -2,48 +2,32 @@ use core::num::{IntErrorKind, ParseIntError};
 
 use const_panic::{FmtArg, StdWrapper};
 
-macro_rules! test_val {
-    ($value:expr) => ({
-        let val = $value;
-        let display = format!("{}", val);
-        let debug = format!("{:?}", val);
-        assert_eq!(
-            trunc_fmt!(1024; StdWrapper(&val).to_panicvals(FmtArg::DEBUG)),
-            &*debug,
-        );
-        assert_eq!(
-            trunc_fmt!(1024; StdWrapper(&val).to_panicvals(FmtArg::DISPLAY)),
-            &*display,
-        );
-    })
-}
-
 #[test]
 fn test_parse_int_error() {
     {
         let err: ParseIntError = u32::from_str_radix("", 10).unwrap_err();
         assert_eq!(*err.kind(), IntErrorKind::Empty);
-        test_val! {err}
+        test_val! {err, no_alternate}
     }
     {
         let err: ParseIntError = u32::from_str_radix("A", 10).unwrap_err();
         assert_eq!(*err.kind(), IntErrorKind::InvalidDigit);
-        test_val! {err}
+        test_val! {err, no_alternate}
     }
     {
         let err: ParseIntError = u8::from_str_radix("256", 10).unwrap_err();
         assert_eq!(*err.kind(), IntErrorKind::PosOverflow);
-        test_val! {err}
+        test_val! {err, no_alternate}
     }
     {
         let err: ParseIntError = i8::from_str_radix("-256", 10).unwrap_err();
         assert_eq!(*err.kind(), IntErrorKind::NegOverflow);
-        test_val! {err}
+        test_val! {err, no_alternate}
     }
     {
         let err: ParseIntError = "0".parse::<core::num::NonZeroI8>().unwrap_err();
         assert_eq!(*err.kind(), IntErrorKind::Zero);
-        test_val! {err}
+        test_val! {err, no_alternate}
     }
 }